pyproject.toml
[project]
name = "agent-exploration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
